refactor & simplify II
