+++ crates/sncast/src/starknet_commands/config_path.rs
@@ -1,5 +1,9 @@
+use anyhow::Result;
 use clap::Args;
-use sncast::helpers::config::ConfigFilePaths;
+use configuration::find_config_file;
+use foundry_ui::components::warning::WarningMessage;
+use sncast::helpers::config::get_or_create_global_config_path;
+use sncast::response::ui::UI;
 use sncast::response::config_path::ConfigPathResponse;
 
 #[derive(Args)]
@@ -9,10 +13,24 @@ use sncast::response::config_path::ConfigPathResponse;
 )]
 pub struct ConfigPath {}
 
-#[must_use]
-pub fn config_path(paths: ConfigFilePaths) -> ConfigPathResponse {
-    ConfigPathResponse {
-        local_config: paths.local,
-        global_config: paths.global,
-    }
+pub fn config_path(ui: &UI) -> Result<ConfigPathResponse> {
+    let local = find_config_file()
+        .ok()
+        .map(|path| path.canonicalize_utf8().unwrap_or(path));
+
+    // TODO: consider adding a wrapper-helper for this logic
+    let global = match get_or_create_global_config_path() {
+        Ok(path) => Some(path.canonicalize_utf8().unwrap_or(path)),
+        Err(err) => {
+            ui.print_warning(WarningMessage::new(format!(
+                "Could not get or create global config file: {err:?}. Proceeding without global config."
+            )));
+            None
+        }
+    };
+
+    Ok(ConfigPathResponse {
+        local_config: local,
+        global_config: global,
+    })
 }
